// packages/ui-cards/src/boards/types.ts
export interface IOptions {
  EditForm: any;
  Item: any;
  type: string;
  title: string;
  queriesName: {
    itemsQuery: string;
    itemsTotalCountQuery: string;
    detailQuery: string;
    archivedItemsQuery: string;
    archivedItemsCountQuery: string;
  };
  mutationsName: {
    addMutation: string;
    editMutation: string;
    removeMutation: string;
    changeMutation: string;
    watchMutation: string;
    copyMutation: string;
    archiveMutation: string;
  };
  queries: {
    itemsQuery: string;
    itemsTotalCountQuery: string;
    detailQuery: string;
    archivedItemsQuery: string;
    archivedItemsCountQuery: string;
  };
  mutations: {
    addMutation: string;
    editMutation: string;
    removeMutation: string;
    changeMutation: string;
    watchMutation: string;
    archiveMutation: string;
    copyMutation: string;
  };
  texts: {
    addText: string;
    addSuccessText?: string;
    updateSuccessText: string;
    deleteSuccessText: string;
    copySuccessText: string;
    changeSuccessText: string;
  };
  isMove: boolean;
  getExtraParams: (queryParams: any) => any;
}

export interface IPipeline {
  _id: string;
  name: string;
  boardId: string;
  tagId?: string;
  visibility: string;
  status: string;
  createdAt: Date;
  createdUser: IUser;
  members?: any[];
  departmentIds?: string[];
  memberIds?: string[];
  condition?: string;
  label?: string;
  bgColor?: string;
  isWatched: boolean;
  startDate?: Date;
  endDate?: Date;
  metric?: string;
  hackScoringType?: string;
  templateId?: string;
  state?: string;
  itemsTotalCount?: number;
  isCheckDate?: boolean;
  isCheckUser?: boolean;
  isCheckDepartment?: boolean;
  excludeCheckUserIds?: string[];
  numberConfig?: string;
  numberSize?: string;
}
// ...
export interface IStage {
  _id: string;
  name: string;
  type: string;
  probability: string;
  index?: number;
  itemId?: string;
  unUsedAmount?: any;
  amount?: any;
  itemsTotalCount: number;
  formId: string;
  pipelineId: string;
  visibility: string;
  memberIds: string[];
  canMoveMemberIds?: string[];
  canEditMemberIds?: string[];
  departmentIds: string[];
  status: string;
  order: number;
  code?: string;
  age?: number;
  defaultTick?: boolean;
}
// ...
export type StagesQueryResponse = {
  stages: IStage[];
  loading: boolean;
  refetch: ({ pipelineId }: { pipelineId?: string }) => Promise<any>;
};
// ...
export type Item = {
  name: string;
  color: string;
};